File Upload and download module done

+++ controllers/admin.js
@@ -1,6 +1,8 @@
 const Product = require("../models/product");
 const { validationResult } = require("express-validator/check");
 
+const fileHelper = require("../util/fileHelper");
+
 exports.getAddProducts = (req, res, next) => {
   res.render("admin/edit-product", {
     pageTitle: "Add Product",
@@ -14,21 +16,20 @@ exports.getAddProducts = (req, res, next) => {
 
 exports.postAddProducts = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const imageUrl = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const userId = req.user._id;
   const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res.render("admin/edit-product", {
+  if (!errors.isEmpty() || !imageUrl) {
+    return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "admin/add-product",
       editing: false,
       hasError: true,
-      errorMessage: errors.array()[0].msg,
+      errorMessage: !imageUrl ? "add proper Image" : errors.array()[0].msg,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description,
       },
@@ -37,7 +38,7 @@ exports.postAddProducts = (req, res, next) => {
   const product = new Product({
     title: title,
     price: price,
-    imageUrl: imageUrl,
+    imageUrl: "/" + imageUrl.path,
     description: description,
     userId: userId,
   });
@@ -47,6 +48,7 @@ exports.postAddProducts = (req, res, next) => {
       res.redirect("/");
     })
     .catch((err) => {
+      console.log("check");
       const error = new Error(err);
       error.httpsStatusCode = 500;
       return next(error);
@@ -122,7 +124,7 @@ exports.getEditProducts = (req, res, next) => {
 exports.postEditProducts = (req, res, next) => {
   const id = req.body.productId;
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const imageUrl = req.file;
   const price = req.body.price;
   const description = req.body.description;
   // const userId = req.user._id;
@@ -134,10 +136,9 @@ exports.postEditProducts = (req, res, next) => {
       path: "admin/add-product",
       editing: true,
       hasError: true,
-      errorMessage: errors.array()[0].msg,
+      errorMessage: !imageUrl ? "add proper image" : errors.array()[0].msg,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description,
       },
@@ -151,7 +152,10 @@ exports.postEditProducts = (req, res, next) => {
       }
       product.title = title;
       product.price = price;
-      product.imageUrl = imageUrl;
+      if (imageUrl) {
+        fileHelper.deleteFile(imageUrl.substring(1, imageUrl.length + 1));
+        product.imageUrl = "/" + imageUrl.path;
+      }
       product.description = description;
       return product.save().then(() => {
         res.redirect("/");
@@ -219,7 +223,15 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProducts = (req, res, next) => {
   const productId = req.body.productId;
-  Product.deleteOne({ _id: productId, userId: req.user._id })
+  Product.findById(productId)
+    .then((product) => {
+      if (!product) return next(new Error());
+
+      fileHelper.deleteFile(
+        product.imageUrl.substring(1, product.imageUrl.length + 1)
+      );
+      return Product.deleteOne({ _id: productId, userId: req.user._id });
+    })
     .then(() => {
       res.redirect("/admin/products");
       console.log("Product deleted");
